Update `nix` dependency to fix compatibility warning for `bitflags`
When compiling Sandcrust on a recent Rust compiler (rustc
1.89.0-nightly), the following warning is issued:

    warning: the following packages contain code that will be rejected
    by a future version of Rust: bitflags v0.7.0

This ancient version of `bitflags` is a depency of `nix`, which was
pinned to v0.8.x. This commit updates `nix` to v0.30.1, which pulls in
a more recent `bitflags` that will continue to be supported by newer
versions of the Rust compiler.

It changes the API around file descriptors slightly, and makes the
`fork` call unsafe. This commit also includes changes necessary to be
compatible with these new APIs.

There's a potential question concerning soundness with `fork` being
made unsafe. `nix`'s safety docs state that:

    In a multithreaded program, only async-signal-safe functions like
    pause and _exit may be called by the child (the parent isn't
    restricted). Note that memory allocation may not be
    async-signal-safe and thus must be prevented.

I don't think that Sandcrust currently includes mechanisms to enforce
the above for child processes. On the other hand, this is a not a new
soundness issue, just one that was raised by the change in `nix`. So I
don't think this demands fixing as part of this commit.